refactor: standardize arcade examples on standalone test module layout
